+++ src/components/Avatar/tests/Avatar.stories.tsx
@@ -1,4 +1,6 @@
+import { useState } from "react";
 import { StoryObj } from "@storybook/react";
+import { expect, fn, Mock, waitFor } from "@storybook/test";
 import { Example } from "utilities/storybook";
 import Avatar from "components/Avatar";
 import View from "components/View";
@@ -15,18 +17,40 @@ export default {
 };
 
 export const src: StoryObj = {
-	name: "src, initials, icon",
+	name: "src, alt",
 	render: () => (
 		<Example>
-			<Example.Item title="With image">
-				<Avatar src="https://pbs.twimg.com/profile_images/1096029593335676929/OZbE9ZXV_400x400.png" />
+			<Example.Item title="src, alt">
+				<Avatar
+					src="https://images.unsplash.com/photo-1536880756060-98a6a140f0a7?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1600&q=80"
+					alt="Amsterdam canal"
+				/>
 			</Example.Item>
 
-			<Example.Item title="With initials">
+			<Example.Item title="src">
+				<Avatar src="https://images.unsplash.com/photo-1536880756060-98a6a140f0a7?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1600&q=80" />
+			</Example.Item>
+		</Example>
+	),
+	play: ({ canvas }) => {
+		const presentation = canvas.getByRole("presentation");
+		const img = canvas.getByRole("img");
+
+		expect(presentation).toBeInTheDocument();
+		expect(img).toBeInTheDocument();
+		expect(img).toHaveAccessibleName("Amsterdam canal");
+	},
+};
+
+export const initials: StoryObj = {
+	name: "initials, icon",
+	render: () => (
+		<Example>
+			<Example.Item title="initials">
 				<Avatar initials="RS" />
 			</Example.Item>
 
-			<Example.Item title="With icon">
+			<Example.Item title="icon">
 				<Avatar icon={IconZap} />
 			</Example.Item>
 		</Example>
@@ -75,7 +99,7 @@ export const squared = {
 			<Example.Item title="squared, with image">
 				<Avatar
 					squared
-					src="https://pbs.twimg.com/profile_images/1096029593335676929/OZbE9ZXV_400x400.png"
+					src="https://images.unsplash.com/photo-1536880756060-98a6a140f0a7?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1600&q=80"
 				/>
 			</Example.Item>
 			<Example.Item title="squared, with initials">
@@ -89,39 +113,92 @@ export const squared = {
 };
 
 export const colors = {
-	name: "color",
+	name: "color, variant",
 	render: () => (
 		<Example>
 			<Example.Item title="color: neutral">
-				<Avatar color="neutral" icon={IconZap} />
-			</Example.Item>
-			<Example.Item title="color: neutral, variant: faded">
-				<Avatar color="neutral" variant="faded" icon={IconZap} />
+				<View gap={4} direction="row">
+					<Avatar color="neutral" icon={IconZap} />
+					<Avatar color="neutral" variant="faded" icon={IconZap} />
+				</View>
 			</Example.Item>
 			<Example.Item title="color: primary">
-				<Avatar color="primary" icon={IconZap} />
-			</Example.Item>
-			<Example.Item title="color: primary, variant: faded">
-				<Avatar color="primary" variant="faded" icon={IconZap} />
+				<View gap={4} direction="row">
+					<Avatar color="primary" icon={IconZap} />
+					<Avatar color="primary" variant="faded" icon={IconZap} />
+				</View>
 			</Example.Item>
 			<Example.Item title="color: positive">
-				<Avatar color="positive" icon={IconZap} />
-			</Example.Item>
-			<Example.Item title="color: positive, variant: faded">
-				<Avatar color="positive" variant="faded" icon={IconZap} />
+				<View gap={4} direction="row">
+					<Avatar color="positive" icon={IconZap} />
+					<Avatar color="positive" variant="faded" icon={IconZap} />
+				</View>
 			</Example.Item>
 			<Example.Item title="color: warning">
-				<Avatar color="warning" icon={IconZap} />
-			</Example.Item>
-			<Example.Item title="color: warning, variant: faded">
-				<Avatar color="warning" variant="faded" icon={IconZap} />
+				<View gap={4} direction="row">
+					<Avatar color="warning" icon={IconZap} />
+					<Avatar color="warning" variant="faded" icon={IconZap} />
+				</View>
 			</Example.Item>
 			<Example.Item title="color: critical">
-				<Avatar color="critical" icon={IconZap} />
-			</Example.Item>
-			<Example.Item title="color: critical, variant: faded">
-				<Avatar color="critical" variant="faded" icon={IconZap} />
+				<View gap={4} direction="row">
+					<Avatar color="critical" icon={IconZap} />
+					<Avatar color="critical" variant="faded" icon={IconZap} />
+				</View>
 			</Example.Item>
 		</Example>
 	),
+};
+
+export const fallback: StoryObj<{ handleError: Mock }> = {
+	name: "test: fallback",
+	args: {
+		handleError: fn(),
+	},
+	render: (args) => {
+		const [error, setError] = useState(false);
+
+		return (
+			<Avatar
+				src={error ? undefined : "/foo"}
+				icon={IconZap}
+				imageAttributes={{
+					onError: () => {
+						setError(true);
+						args.handleError();
+					},
+				}}
+			/>
+		);
+	},
+	play: async ({ args }) => {
+		await waitFor(() => {
+			expect(args.handleError).toHaveBeenCalledTimes(1);
+		});
+	},
+};
+
+export const className: StoryObj = {
+	name: "className, attributes, imageAttributes",
+	render: () => (
+		<div data-testid="root">
+			<Avatar
+				src="https://images.unsplash.com/photo-1536880756060-98a6a140f0a7?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1600&q=80"
+				className="test-classname"
+				attributes={{ id: "test-id" }}
+				imageAttributes={{ id: "test-image-id", alt: "test image" }}
+			/>
+		</div>
+	),
+	play: async ({ canvas }) => {
+		const root = canvas.getByTestId("root").firstChild;
+		const img = canvas.getByRole("img");
+
+		expect(root).toHaveClass("test-classname");
+		expect(root).toHaveAttribute("id", "test-id");
+
+		expect(img).toHaveAttribute("id", "test-image-id");
+		// Uses alt attribute in case prop is not passed
+		expect(img).toHaveAccessibleName("test image");
+	},
 };