+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - useHandlerRef: new hook
 - TrapFocus: new utility, new trap mode
 - Icon: Added vertical alignment in case svg content is smaller that the icon size
+- Avatar: Image attributes, relaxed types, more tests